- Aligned parameter order of EnvelopeNet constructor with Envelope - Removed NumParts/NumPaths/NumPoints properties, replaced with properties to automatically determine from array length - Added MP/Poly/Line header structs for ease of reading/writing - Added support for writing Points, Polylines and Polygons (polygons not tested yet) - Added test for editing points/polylines - Added test for creating a new polyline

+++ FileGDB_DotNet/GeometryNet.h
@@ -10,11 +10,11 @@ namespace FileGDB_DotNet
 	public ref class GeometryNet
 	{
 	public:
-		GeometryNet(int shapeType)
+		GeometryNet(long shapeType)
 		{
 			this->shapeType = shapeType;
 		}
 
-		int shapeType;
+		long shapeType;
 	};
 }
